Add “Go to line” functionality through URL
Adding a "#l27" suffix to the URL makes the script scroll to line 27 and
to select this line (other possible syntaxes: "#Line27", "#line27", and
"#L27".

WARNING: this is not implemented with HTML anchors, so editing the URL
from within the address bar and typing “Enter” will not trigger the
scroll. The anchor MUST be part of the URL when the script loads. If a
user adds it afterwards, they have to reload Enlight script for this
page.