Upgrade highlight.js to version 9.2.0
I had to skip some earlier versions of highlight.js because of a code
modification preventing me to call the library − see discussion on this
commit:
https://github.com/isagalaev/highlight.js/commit/d6207b0720e457bdf4646202336c3d2bef6ce642

Kindly fixed by one of the core contributers in this commit:
https://github.com/isagalaev/highlight.js/commit/20346ce322d958cb627e51aa72d3a5eeedb65330

So now we are with an up-to-date version again.
But the plan is to submit the `all-languages` version of the add-on on
AMO, because due to recent changes in Firefox, users will not be able to
install this otherwise unsigned version and to benefit from complete
language support (except for Nightly / Developer Edition users).
If this works, `all-languages` would be merged into `master` branch.