Properly remove trailing blank line produced by line numbering.
This time we only make sure that the end of the string containing the
highlighted code is not a '\n' character.

Previously, when the file would end with a newline, a number would be
added once for the newline character, and then once for the end of the
string. If we remove the end of the string, we have no number on the
last line for files with no ending '\n'. So we add numbers on all lines
ending with a '\n', plus an additional number on the last line if and
only if it is preceded by a character other than a newline (matched by
the dot added in this commit).