Switch to web extension framework
Update all the source code to have a working web extension, compatible
with Firefox 48+? (tested on version 57).

Background script has been completely rewritten, even though much of the
logics remain the same. Content script remains closer to its previous
version. Option handling must now be done by the extension (no more by
the SDK): add an option page. The popup for language selection did not
change much.

Since we changed a lot of things, this commit (or rather, the previous
one) also makes some cleanup in the files tracked by Git. Furthermore,
README.md and icons have been updated.

Of course, this new version of the add-on is no more compatible with the
jpm tool, or with the add-on SDK APIs.

Change version number to 2+. Since the new reference for version numbers
allows for more dot-separated values, the add-on will follow the
following scheme: A.B.X.Y, where
- A is the major version, bumped to 2,
- B is the minor version, incremented for each modification to the
  add-on internal code,
- X is the major version of embedded highlight.js,
- Y is the minor version of embedded highlight.js.
highlight.js also has a third (build) number, that will not appear in
the add-on version number.

Differences in behavior:
- The icon of the button no longer changes aspect when the button is
  clicked.
- Some options disappeared, as I suspect nobody used them: namely, the
  option for preventing background coloring, as well as the option for
  using alternate icon set (now handled directly in the manifest) have
  both been removed.

Known issue: the add-on keeps one global state to remember if the tab
has already been colored or not. Of course, in the case when several
tabs are to be highlighted, this is a problem. This is not so big an
issue, since in this case:
- If the tab should be colored but the state is not correct, the first
  click to the add-on button is without effect. Second click works as
  expected and opens the language selection popup.
- If the tab should be reset (to no-color) but the state is not correct,
  the language selection pops up (instead of not showing), and a click
  will un-highlight tab contents, whatever language was selected. Next
  click will have status in order, and will act as expected.
As a workaround, maybe the background script could probe the tab to see
if content script has been injected already. But I haven't been
successful with this so far, background script seems to hang if it
fails to connect to a content script. The issue will probably stay there
for a while, until I find something better.