Fox a bug for preferences display (checkboxes not checked on page load)
Whether auto-detection and line numbering would be activated or not in
the preferences, the relevant checkboxes would never be checked when the
preferences page loads. This is because we're not assigning the
preference value to the correct attribute ("value" instead of "checked")
for the boxes. Let's fix this.

May be related to issue #9 (not sure).

Fixes: 99821c9abdf9 ("Switch to web extension framework")