Reorder languages in popup in a single alphabetical list
The languages in the popup were sorted in two consecutive alphabetical
list for historical reasons: The first list, containing the 22
“most-often” used languages (as per the bundle provided by
highlight.js), was all I managed to use in the first version of the
add-on to get it published on AMO (because the minified embedded library
is impossible to review, so I need an “official source” to get it from,
and that was complicated).

Anyway. When all the other languages made it to the popup, as a second
list, it made things very difficult to understand for new users, often
not realising that there are two lists. It's probably not worth having
those two lists at all, so let's just merge all languages into a single
list in order to make it easier to find whatever language is desired for
syntax highlighting.

Fixes #10.