Fix report of detected language when auto-hilight is off
When automatic syntax highlighting is deactivated, the listener supposed
to receive the name of the language detected by the content script when
auto-detection is required (i.e. when user manually selects
“auto-detect” in the popup) is never launched, so the tooltip of the
add-on button is not correctly updated with the name of that language.

Fix it by setting up the listener when the user clicks on the button to
open the popup.