Add an option to detect language based on file extenstion (from URL)
This is for when the script attempts to “auto-detect”, and instead of
relying on highlight.js to perform auto-detection.

Pros:
- So much faster, especially for big files.
- Probably more accurate, since highlight.js tries to guess among so
  many different languages.
- At last we avoid highlighting text files (.txt).

Cons:
- Many files with extensions won't show up in Firefox, which proposes to
  download them instead.
- There is some collision on the file extensions, for example: .sh can
  be bash or shell.

I just implemented detection for a few of the most often used languages.
We can add others if users request them. If the add-on does not know the
extension, it falls back on syntax-based detection from highlight.js.

Because this is an important change with the previous behaviour, add an
opt-in option for using this detection method instead of the
syntax-based way.

Fixes #11.