Switch to highlight.js v9.14.2. Enlight v2.3.9.14.
Highlight.js is active again, let's update the lib. This adds 9 new
languages, including a “Plain text“ which is in fact an alias to the
"nohighlight" class telling the lib to not color the text. Let's add it
to the list anyway for consistency. There are 10 new color schemes.

Bump the version number of the add-on accordingly.