Move JSON file with languages from popup/ to options/
In preparation for adding an option for selecting a subset of languages
for the popup and for automatic detection, move the JSON file containing
the list of supporting languages to the options/ directory, where it
will be used for generating a list in preferences. It is expected that
the popup and background script will cease to use the JSON file anyway
once we have a preference that indicates what languages to use (or maybe
not actually, we may still need it to retrieve the human-friendly names
for the languages).