Save and restore options for the subset of languages to use
We recently introduced a menu to select the languages to support for
autodetection (and to display in the popup), but the preferences is not
saved or restored. Implement save/restore.

Also move default options to a new separate file, called earlier in the
HTML menu for preferences, so we can reuse it in other script files.

The default list of languages is set to what highlight.js suggests by
default on its download page, https://highlightjs.org/download/