Use the selected subset of languages for autodetection
In order to improve the performance and the accuracy of autodetection,
highlight.js offers the possibility to work only with a subset of all
supported languages. Now that we have a user preference to select such a
subset, let's use it to limit the languages to use at runtime.

We add a duplicate list of default languages, which I do not really
like, but I do not see another solution without parsing a JSON file,
which I prefer to avoid.