Switch to highlight.js v10.1.1. Enlight v2.5.10.1.
The highlight.js library introduced some breaking changes in version
v10.0, but thankfully nothing that changes the mechanics we use. A
couple of languages or stylesheets were renamed. One stylesheet alias
was removed. See [0] for details. Let's update the lists and take this
opportunity to re-align languages names on the list on the highlight.js
download page.

The behaviour of the add-on itself is unchanged, with a minor exception:
the default set of languages used for auto-detection is realigned on the
list of "common" languages from highlight.js download page, meaning that
it grew (13 more entries).

Do not offer "C-like" language, which seems to be only a dependency
included in C/C++ languages.

Also improve file-extension-based detection accordingly.

[0] https://github.com/highlightjs/highlight.js/blob/master/VERSION_10_BREAKING_CHANGES.md