Switch to highlight.js v11.0.1
This new highlight.js version comes with breaking changes, and the
add-on needs to be adapted somewhat. In particular:

- The class name applied to the block to highlight changes ("hljs" is
  not longer needed, instead the language name is prefixed with
  "language-").
- The function that we used previously has been deprecated, let's use
  the new one.
- The supported languages, and mostly the styles, have changed and need
  to be updated. In particular, a number of style sheets have been
  moved to a base16/ directory.

Note also that the way we retrieve the library has changed. Previously,
we would download a generated archive from the download page of the
hightlight.js website, after activating the checkboxes for all
languages. This archive would contain the style sheets and a minified
script with all supported languages. But the way the library is packaged
has changed, and the script no longer comes with direct support for all
languages. Instead, they have to be programmatically "registered" at
runtime, to avoid loading too many language definitions. So instead we
build the script with support for all languages ourselves, directly from
the sources of the project.

We could consider alternative approaches in the future (load the library
from the content script as a dynamic module, and import support for the
languages selected in the options?). No immediate work on that side.

This commit also contains some additional miscellaneous, minor
improvements to the code.

The add-on version is updated in the manifest to reflect the new version
number for the embedded library, but this will not be a tagged release
given that more changes are following.