Fix popup closing (wait for the promise to return)
The popup code closes the popup window when a user selects a language to
use for highlighting. It does so right after sending a message to the
background script to tell it about the selected message. It would not
wait for the related promise to return, making Thunderbird complain in
the console. Let's fix it.