Switch to a better JSON structure for the list of languages
The list of languages supported by highlight.js would be stored as a
JSON object for which the keys would be the language ids, and the value
their names. It feels more logical to have instead an array of
languages, each with an id ("class", since they are class names in the
content script) and a name.

We need an additional loop when looking for the name corresponding to a
given id (to update the title of the add-on button), but overall the
code looks cleaner and we do not have to loop on
"Object.keys(languageList)" anymore.