Fix race condition between language list and checkboxes (in the options)
The preferences page has a list with all languages supported by
highlight.js, and the user can pick those they desire to use with the
add-on. This list is generated programmatically from a JSON object.
Furthermore, the selected languages are retrieved from user preferences,
and the related checkboxes are activated accordingly to reflect those
currently in use.

There is a race condition here, because we would not wait for the list
to be fully generated before trying to enable the checkboxes. This may
end up with all checkboxes being unchecked. This is an issue by itself,
but worse than that, any subsequent modification of the options triggers
a save of all options in the page, including of the empty list of
languages to use, leading to no language used and an empty popup when
trying to highlight a document.

Make sure that the list is generated before trying to activate the
checkboxes: wait for a dedicated event added for the circumstance.