Switch to JavaScript modules in the add-on code
Turn the JavaScript scipts into modules for the background script and
the preference page. This makes it possible to better share some
elements (such as the default values for the options) between the
scripts, and avoid having to resort to XMLHttpRequest() to load a JSON
file. Instead, we can simply import the necessary definitions.

Some functions and files for the preference page are reworked and
reorganised, to make the code cleaner.