Name const variables with a capital letter
Some variables just happened to be declared as const because they are
never reassigned. Others are constants of the more traditional sort:
they represent a fixed value that we do not want to hardcode, global to
the file. Instead of prefixing them with a "g", let's just give them a
capital letter to clearly mark them as such.