Rework the aspect of the preferences page
Improve help messages and overall clarity and style for the preferences
pages.

I am proud to have figured out a way to pull the browser style directly
from the chrome stylesheet, and to have found the few necessary
adjustements to "body" to have it work correctly and adapt to the
browser's style (be it with a light or a dark theme). Next time: see if
the same thing works for Thunderbird...

No functional change in the available options.