Add .eslintrc.json for ESLint
Let's configure a JavaScript linter for the code. We will use ESLint
(https://eslint.org/), and we add a dedicated configuration file.

Use "JavaScript Standard Style" (https://github.com/standard/standard)
with a few modifications:

- Allow dangling commas for some constructs (It avoids touching
  unnecessary lines in Git history when adding new items to arrays or
  objects).
- Reject blocks in curly braces on a single line (Personal preference).
- Use four spaces for indent (I find I don't like having two spaces
  only).
- Use semicolons at the end of statements (To avoid destroying all the
  Git history).
- Use double quotes for strings (Idem).
- Allow multiline strings, we have one and we are one hundred percent
  sure it will be supported in our context.

To run the linter, install ESLint and the "standard" package:

    $ npm install eslint --save-dev
    $ npm install standard --save-dev

DO NOT RUN: Config file initialisation (but we add one here already):

    $ npx eslint --init

Then run the linter on the files (or directories)

    $ npx eslint scripts options popup

Alternatives such as yarn are available. For more details, see:
https://eslint.org/docs/user-guide/getting-started