Rename content and background scripts
The main scripts of the add-on (content and background scripts) have
their names prefixed with "enlight-". There is no real justification to
that other than to highlight that they implement the logics for the
add-on. But there's no real reason to use the name of the add-on itself
in the filenames (the root directory, the add-on title, the README.md
will already tell you how the add-on is named). Plus it makes file names
needlessly longer to type, and does not help with file name completion.
Let's rename the scripts by removing this prefix.