Add highlight.js as a Git submodule
Instead of cloning the Git repository manually from the Makefile, let's
directly add it as a Git submodule to the project. We do not have a
shallow repository anymore, but we can change to a newer version
without having deleting and cloning again the repository. It looks
cleaner, and better mark highlight.js as a direct dependency to build
the add-on.