Rework detected language retrieval in content script
We previously assumed that the class generated for the <code> element
was of the shape "hljs language-<id>". This is correct for
autodetection, but this is wrong when we use the file extension to
find out the language. In that case, the two names are swapped,
"language-<id> hljs". So we cannot simply remove the prefix, we need to
split and find the correct part of the className.